Superty_mall/requirements.txt
# Usage: pip install -r requirements.txt -i https://pypi.tuna.tsinghua.edu.cn/simple some-package

Django==1.11.11
django-crontab==0.7.1

django-redis==4.10.0
elasticsearch==2.4.1
Pillow==6.2.1
itsdangerous==1.1.0
PyMySQL==0.9.3
redis==3.0.1

amqp==2.4.0
appnope==0.1.0
backcall==0.1.0
billiard==3.5.0.5
celery==4.2.1
certifi==2018.11.29
chardet==3.0.4
decorator==4.3.2
dnspython==1.16.0
eventlet==0.24.1     
# fdfs-client-py==1.2.6
greenlet==0.4.15     
idna==2.8        
ipython==7.3.0      
ipython-genutils==0.2.0
jedi==0.13.2     
Jinja2==2.10       
kombu==4.2.2.post1
MarkupSafe==1.1.0      
monotonic==1.5
mutagen==1.42.0
parso==0.3.4
pexpect==4.6.0      
pickleshare==0.7.5
pika==0.13.0
# pip==19.0.3
prompt-toolkit==2.0.9      
ptyprocess==0.6.0
pycryptodomex==3.7.2
Pygments==2.3.1
python-alipay-sdk==1.10.0
pytz==2018.9
QQLoginTool==0.3.0
requests==2.21.0     
setuptools==40.8.0
six==1.12.0
traitlets==4.3.2      
urllib3==1.24.1
vine==1.2.0
wcwidth==0.1.7
wheel==0.33.1
# django-haystack==2.8.1